requirements.txt
# Core Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI
anthropic>=0.25.0

# Cache
redis>=5.0.3

# HTTP Client
httpx>=0.27.0

# Payments
stripe>=8.0.0

# Production
slowapi>=0.1.9
python-multipart>=0.0.9
